test(backup/gcs): use separate class for config integration tests
